refactor(storage): move `ProjectTeam` to a public header
